requirements.txt
discord.py
asyncpg
asyncio
datetime
pytz
asqlite